rabbitmq: Use N/2+1 queue mirrors instead of N.
Performance testing has demonstrated that using N/2+1
queue mirrors instead of N improves maximum throughput
by ~40% and RPC call latency by 60% in a 3-node setup.
It is expected that such benefits would continue to
pile on when provisioning additional RMQ nodes.

Using slightly fewer queue mirrors should be safe as
we currently use a pause_minority partition/recovery
strategy, meaning that the cluster should pause before
or when all replicas of a queue are lost.

The only downside to this is that when a cluster member
drops, there may be some intermittent pausing as mirrors
are synchronized to account for the change in cluster
membership. This itself can be countered by tuning the
ha-sync-batch-size setting.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>