chef: Change ownership/mode of conf files
This commit ensures that automation renders all OpenStack
service configuration files (".conf") with mode 0640 and
sets user:group ownership as root:<roleaccount>.

In principle, from a security perspective, this is "better"
than what we are currently doing, which is to use mode 0600
and <roleaccount>:root.  Role accounts should not be writing
to configuration; they should only consume it.

This also alleviates some slight aches and pains with
in-place OpenStack upgrades, as Ubuntu packaging appears to
finagle with the modes of the configuration files.  Working
in the same direction that it finagles things results in
the headnodes being closer to convergence after the package
upgrade operations run.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>