nova-compute: Avoid rendering of invalid config
If cpu_mode is _host-(model|passthrough)_, then cpu_model
should be disregarded; otherwise, one would have requested
a _custom_ CPU model or not specified to use the hosts'
configuration.

In fact, `nova` throws an exception during scheduling if
the `nova-compute` service specifies cpu_model when cpu_mode
is not _custom_, or when a cpu_model is not specified, but
cpu_mode is _custom_.

https://github.com/openstack/nova/blob/stable/rocky/nova/virt/libvirt/driver.py#L3983
```
Invalid: A CPU model name should not be set when a host CPU
model is requested
```

Thus, even if the chef-databag specifies this invalid
set of parameters, this commit ensures that the resulting
config will not result in scheduling errors.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>