create-virtual-*: Parallelize spinups/reboots
We previously used `vagrant reload` to turn all the VMs
one-by-one after they were provisioned with `vagrant up`.
Unfortunately, `vagrant reload` has no `--parallel` flag
as does `vagrant up`.

However, we have Ansible, and we generate an Ansible
playbook as part of the create-virtual-environment.sh
script, so just leverage that to do the reboots instead,
and do so in parallel.

Moreover, add `--parallel` to all `vagrant up` commands
to allow provisioners which respect this flag to come up
faster.

This improves build time of a 1h1w by ~1.8% in my lab.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>