mysql: Wait for Consul to elect a mysql primary
The MySQL recipe does not wait for Consul to elect a
database primary, and the next recipe to currently execute
(powerdns) tries to make use of the database host.  This
causes a race condition that has now begun to enumerate
more frequently... so wait for Consul to elect the primary
MySQL host as part of the MySQL recipe.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>