ansible: Default to HWE kernel configuration
We now install the HWE kernel by default as part of the
packer/virtual builds, so we should make this the default
in Ansible.

Failing to do so results in kernel-utilities, such as
`bpftool`, `perf`, etc. being unavailable.  This issue has
manifested itself already as a Calico issue, as
`calico-felix` needs `bpftool` but it was not being
installed.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>