configure-chrony: fix IPv4/6 compatibility issue
When the chef-bcpc `system_enable_ipv6` attribute is set to
`False`, it may be the case that `chrony` still tries to
open `AF_INET6` sockets to connect to timeservers.

This will result in `chrony` failing to start with the
following logged to the system journal:
```
Jun 08 21:37:38 <hostname> chronyd[5708]: Could not open IPv6 command socket : Address family not supported by protocol
```

To avoid this, configure `chrony` to start with `-4` when
`system_enable_ipv6` is set to `False`.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>