plug_tap: Do not request MQ tuntaps if using 1vCPU
Multiqueue is applied by tagging an image with a property,
denoting that all instances spun off that image should use
multi-queue.  However, in the case of flavors with a single
vCPU, multi-queue makes no sense.

This commit fixes the case where instance creation with
a multi-queue tagged image fails when the underlying
instance's flavor specifies a single vCPU.  Prior to this
commit, a multi-queue tuntap would be created, but libvirt
would expect non-multi-queue tuntap and would fail to start
the instance.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>