packer: Install tools mated with kernel
Calico requires linux-tools-`uname -r` to be present.  With
upstream builds, it may be the case that a packer image is
generated on a specific date, with kernel x.y.z installed.
The Ansible automation will install the linux-tools-*
metapackage, which may install kernel x.y.theta tools.

To avoid this drift, install the tools mated with the
kernel at packer provisioning time.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>