ceph: Increase mgr beacon timeout/work periods
The ceph-mgr is not highly threaded (at least in Mimic) and
can become a SPOF for a large Ceph cluster.  Thus, for
large clusters, it may be a good idea to both increase the
beacon timeout and period of cyclic work.  This should
result in the mgr being less susceptible to being presumed
failed during intensive operations on large clusters.

This has the side-effect of reducing the period at which
stats are rendered for commands such as `ceph -s`.

Likewise, enable the number of mon threads to be increased
to provide more headroom during compaction and other highly
CPU-intensive mon operations.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>