Patch an outstanding python3 issue in etcd3gw
etcd3gw can perform a double-decode of a byte sequence
in its watch thread: once via requests/the decode_unicode
parameter, and once itself (explicitly) via the decode
method on the first line of the loop, here:
https://opendev.org/openstack/etcd3gw/src/tag/1.0.0/etcd3gw/watch.py#L22

In python2, this works fine:
```
>>> b'hello_world'.decode('utf-8').decode('utf-8')
u'hello_world'
```

However, in python3, an AttributeError is raised:
```
>>> b'hello_world'.decode('utf-8').decode('utf-8')
Traceback (most recent call last):
File "<stdin>", line 1, in <module>
AttributeError: 'str' object has no attribute 'decode'
```

Incidentally, the Exception raised as a result of this
operation is never materialized due to the fact that
the future (generated on behalf of the ThreadPoolExecutor)
containing the result is never waited on/evaluated, making
it hard to chase down.

This commit patches the issue within this repository until
we have the time to get the fix properly upstreamed.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>