neutron: Pick up MTU settings of the host
Currently, Ansible configures the MTU of the transit links
via the networking['mtu'] variable.

Neutron has a similar-ish configuration tunable for this,
which is `global_physnet_mtu` (default 1500, the standard
for ethernet networks).

The implication of this variable is that one cannot set the
MTU of an OpenStack network more than `global_physnet_mtu`.
And, moreover, all networks created inherent this MTU by
default.

Add plumbing to BCPC such that such that we set the Neutron
tunable to the same value that Ansible configures the host
with. Also add a Chef attribute that retains a 1500 MTU for
newly created networks so that we do not change the default
yet, and allow operators to create networks by default with
a 'standard' MTU even when the host supports jumbo frames.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>