Dynamically set Ceph pool size
Previously, the Ceph pool size was defined statically (to
1) in the Chef attributes. However, this value could not
be set to anything other than 1 without breaking a build
that had less than 1 node with a storagenode type role.

At the same time, building clusters with a size of 1 (as
of Ceph Octopus, or BCC 8.2.0) would result in 3h3w
clusters being stuck in HEALTH_WARN because no replication
was configured/leveraged.

To address these issues, this commit repurposes Ansible's
`reweight-ceph-osds` tag into `configure-ceph`. As part of
this tag, an appropriate CRUSH replication factor is
selected and configured based on the cluster topology and
the specified Chef attribute.

In the future, we might also want to use this tag to
configure the CRUSH replication policy to distribute
objects across racks rather than across hosts by default.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>