Add serial console to default command line
Most virtual environments are spun up off the base packer
image and then never rebooted - even though this option can
be specified in the Chef environment, it is not necessarily
leveraged in such cases.

At the same time, having the serial console aids developers
in troubleshooting virtual builds when the network is
problematic -- so add it as part of the packer build
scripts.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>