Add feature to assist migration of `ceph-mons`
This commit adds a feature to the Nova migration code to
overwrite `ceph-mon` addresses in the domain's XML
specification with those specified in the general section
of `ceph.conf`.

It does the same for newly created and refreshed Nova
block device mappings and Cinder attachments as well.

Together, this assists the IP address migration of `ceph-mon`
nodes, especially for VMs which have a long uptime, by
allowing an operator to reconfigure the set of `ceph-mon`
IPs via a single live-migration of each VM in the cluster.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>