nova-compute: Fix for live-migration issue
qemu suffers from a race condition that bubbles up to
libvirt, and subsequently nova-compute. Specifically,
when this race condition is observed, libvirt raises
an (unhandled) exception that causes nova-compute to
leave shrapnel lying around, and the VM to enter a mostly
undefined state.

Simply retrying the failing libvirt invocation is enough
to workaround the race condition, effectively, based on a
reproducer that we have running locally - which is exactly
what this commit does.

If libvirt appears stuck in this condition (something we
have not seen with our reproducers), assume that the
migration completes to allow nova-compute to continue
making progress and cleaning up the domain.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>