libvirt: Use a dedicated I/O thread for RBDs
qemu's main event loop is ordinarily responsible for
managing all of the emulated devices within an instance.
It can become a bottleneck for VMs with lots of vCPUs
and emulated devices. This event loop is also associated
with a global mutex, because `qemu`'s event handling code
was ordinarily not thread-safe.

To avoid I/O jitter and latency spikes that may be imposed
by the event loop, we can move `librbd` and I/O processing
into dedicated I/O threads, and also optionally pin those
dedicated I/O threads to pCPUs on the hypervisor (in the
same way Nova allows us to pin emulator threads).

This commit is an experimental first stab at implementing
this by unconditionally leveraging a dedicated I/O thread
that is shared by all RBDs attached to a VM. With this
patch, any VMs built new or live-migrated begin using a
dedicated I/O thread for RBD operations.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>