nova: Retain MTU during live-migration
There's a longstanding issue with Nova and Neutron
disagreeing about how the MTU is handled during live-
migration.

In `nova`'s mind (and really, `libvirt`'s mind), the MTU
cannot change during the lifetime of a running VM. It
*must* be stopped in order for an MTU change to be
observed.

However, in `neutron`'s mind, the MTUs should "immediately"
track that of the network and adapt. This is because in
a traditional, non-Calico OpenStack cloud, the MTUs on the
ports and other non-host elements of the network must also
reflect how the hosts are configured -- this is especially
important in a L2 network.

Currently, live-migrations follow the Neutron mentality
wherein a VM's MTU will adapt to the network. But, as per
above, that simply does not work with `nova`/`libvirt`
right now -- an exception is raised during live-migration.

Fortunately, in a Calico OpenStack deployment, we can have
networks with mixed MTUs (as the hosts will fragment as
needed). We just need to "fix" the issue with Nova such
that it continues to allow live-migrations by retaining the
VM's MTU across live-migrations -- which is what this
commit does.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>