nova: Allow mutation of schedulers
Currently, because we use `node.default`, overridden
attributes are ignored. `node.default` seems to be used
because Chef itself does not allow node data structures
to be mutated.

This does not follow the scheme of how we allow things
to be overridden generally, though, so make a copy of
the scheduler filters array when mutating it.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>