horizon: Use same memcached timeouts as elsewhere
When a memcached server becomes unavailable (e.g. as part
of a host reboot), the Horizon dashboard got a little
sluggish. As it turns out, the dashboard is using the
default memcached settings, separate from the tuned ones
that `keystone` is using.

After changing the Horizon memcached timeouts to match
that of `keystone`, the dashboard was a lot more peppy when
one of the memcached instances became unavailable.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>