nova: Prevent misconfiguration of tap MTU
When `nova-compute` is restarted, it calls `plug_vif` on
all VIFs. When this is done, the MTU of the Neutron network
is used.

In cases where instances were spawned prior to the MTU of
the Neutron network MTU changing, this will result in the
tap getting the incorrect MTU. The tap, itself, is used
by `calico-dhcp-agent` to configure `dnsmasq` for the
instances on the host, so it must be set correctly.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>