ceph/quincy: Do not size pools when init'ing cloud
Quincy? has a bug where it flips out about billions of PGs
when trying to set a pool size of 1 and there are no OSDs
which are members of the cluster yet.

So, only create the pools when first standing up the mons
and then revisit it as part of 'configure-ceph'.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>