nova-head: Drop filter before restarting scheduler
There is a possible race condition where the `nova.conf`
has the additional filters we drop in specified in the
configuration (which then fires notification to immediately
restart `nova-scheduler`), but we have not dropped the
module yet. This can raise an exception and prevent an
initial Chef run of `bcpc::nova-head` from completing.

Thus, reorder things within the recipe.

And while we're at it, we might as well move the Ussuri
patches up to where the filters are dropped, so that the
first time `nova-*` is restarted, the code changes take
effect (instead of being deferred until the end of the
initial client run).

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>
Signed-off-by: Piotr Sipika <psipika@bloomberg.net>