rabbitmq: Do not restart the broker late in build
The RabbitMQ broker is configured to not automatically
sync the queues when a broker leaves the cluster at
present. Yet, Chef had multiple deferred reboots and
it was kicking the broker at the end of the client
run.

This results 'weird' corruption/behavior in some versions
of RabbitMQ, and we were probably just flat out lucky
that this was formerly working with 3.8.x for so long.

Do not restart the broker late in the build as we do not
synchronize the queues after!

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>