nova: Fix CPU/RAM scheduler weighers (#2111)
* nova: Fix CPU/RAM scheduler weighers

"Normalize the weights instead of using raw values"
introduced a change whereby all weigher classes' outputs
are normalized to the range [0, 1].

However, the CPU/RAM weighers were not adjusted to use a
percent utilization instead of an absolute value as part
of this change, so the normalization here heavily biases
(stacks) allocations on certain classes of hypervisors
over others initially in a heterogeneous environment.

e.g., suppose we have two hypervisors, one with 512GB RAM
and one with 2TB RAM. If we consider the RAMWeigher in its
current form, it will weigh the 2TB RAM hypervisor >= the
512GB RAM hypervisor until there are 1.5TB of instances on
the former (assume we are looking at RAM in isolation).
Only _then_ will the weigher actually begin giving equal
weights to encourage spreading.

The module-level comment suggests the weigher intends to
spread evenly, but does not define what "evenly" is. It
seems logical that we would want to preferentially inter-
leave across both hypervisors in a biased fashion, rather
than stacking until the resource are at equilbrium and then
spreading after.

This commit makes that change.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>