ceph: Remove defunct settings (#2124)
* ceph: Fix bluestore_cache_size > osd_memory_target

osd_memory_target describes the total amount of memory tha
the entire OSD process should aim for. It makes no sense
to have a particular OSD cache be greater than that value.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>

* ceph: Remove settings which no longer exist

Self-explanatory... these settings are from an older
version of Ceph, and were never removed.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>

* ceph: Remove a former workaround

Octopus 15.2.12 had a bug wherein threaded fscks could
crash/shred the data on the OSD. The issue has long since
been fixed, and we no longer need to disable it.

Parent tracker:
https://tracker.ceph.com/issues/50017

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>

* ceph: Fix an (incorrectly) specified setting

The osd_max_scrubs settings was misdeclared and was not
actually doing anything. The attributes value wanted to
raise it to 5, but we've been using 1 (the default) for
a very long time and have no need to scrub faster.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>

* ceph: Remove unused max_open_files

The max_open_files argument is only used by the init.d
variant of the startup scripts, which have been replaced
by systemd. As this setting has not actually been
respected for quite some time, remove it.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>

* ceph: Reduce osd_memory_target to 6GiB

Now that we are using NUMA-pinned OSD instances, there is:

  a) Less "slack" space that can be shared across all the
     OSD processes, as they are segmented into bins.

  b) More unbalanced use of memory, as the number of OSD
     processes do not evenly divide into the number of
     NUMA nodes on all our platforms.

Thus, we need to drop the osd_memory_target just a bit to
accomodate the new arranagement in order to avoid memory
pressure over time.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>

* ceph config cleanup

Signed-off-by: Jing Geng <jgeng25@bloomberg.net>

Conflicts:
	chef/cookbooks/bcpc/attributes/ceph.rb
	chef/cookbooks/bcpc/templates/default/ceph/ceph.conf.erb

---------

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>
Co-authored-by: Jing Geng <jgeng25@bloomberg.net>